server/requirements.txt
flask==3.0.3
joblib==1.4.2
scikit-learn==1.5.2
numpy==1.26.4
opencv-python==4.10.0.84
google-cloud-vision==3.7.4
protobuf>=3.20.2,<6.0.0
google-generativeai>=0.7.0
python-dotenv >= 1.0.0
transformers
torch
bs4
Pillow
sentence-transformers
pinecone
cryptography
google-cloud-firestore
google-cloud-storage
firebase-admin
dotenv
asyncio
aiohttp
requests
Flask-Limiter
google-cloud-translate
flask-cors
gunicorn
nest_asyncio